add containers dir
