+++ src/App.js
@@ -1,10 +1,10 @@
 import React from "react";
 import { Route, Switch } from "react-router-dom";
 import Header from "./components/Header";
-import Home from "./pages/News/Home";
-import NewsPolitics from "./pages/News/NewsPolitics";
-import NewsTechnology from "./pages/News/NewsTechnology";
-import NewsWorld from "./pages/News/NewsWorld";
+import NewsPage from "./pages/NewsPage";
+// import NewsPolitics from "./pages/News/NewsPolitics";
+// import NewsTechnology from "./pages/News/NewsTechnology";
+// import NewsWorld from "./pages/News/NewsWorld";
 import SearchNewsPage from "./pages/SearchNewsPage";
 import StockListPage from "./pages/StockListPage";
 
@@ -13,10 +13,10 @@ const App = () => {
     <React.Fragment>
       <Header />
       <Switch>
-        <Route exact path="/" component={Home} />
-        <Route path="/news-technology" component={NewsTechnology} />
-        <Route path="/news-politics" component={NewsPolitics} />
-        <Route path="/news-world" component={NewsWorld} />
+        <Route exact path="/" component={<NewsPage category="Business" />} />
+        <Route path="/news-technology" component={<NewsPage category="Politics" />} />
+        <Route path="/news-politics" component={<NewsPage category="ScienceAndTechnology" />} />
+        <Route path="/news-world" component={<NewsPage category="World" />} />
         <Route path="/stock" component={StockListPage} />
         <Route path="/search" component={SearchNewsPage} />
       </Switch>
