+++ src/pages/NewsPage.js
@@ -1,0 +1,14 @@
+import React from "react";
+import News from "../../components/News";
+import ButtonTab from "../../components/ButtonTab";
+
+const NewsPage = ({ category }) => {
+  return (
+    <React.Fragment>
+      <ButtonTab />
+      <News category={category} />
+    </React.Fragment>
+  );
+};
+
+export default NewsPage;
