+++ src/pages/StockListPage.js
@@ -1,11 +1,14 @@
 import React from "react";
 import StockList from "../components/StockList";
 import TextTmplate from "../components/Text";
+import Box from '@material-ui/core/Box';
 
 const StockLists = () => {
   return (
     <React.Fragment>
-      <TextTmplate title="MAJOR STOCK LIST" />
+      <Box m={1}>
+        <TextTmplate title="MAJOR STOCK LIST" />
+      </Box>
       <StockList />
     </React.Fragment>
   );
